requirements.txt

# Core
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
xgboost>=1.7
shap>=0.44
pyyaml>=6.0
matplotlib>=3.7
seaborn>=0.13
